Moved questions into a separate file
