+++ app/sections/FAQ.tsx
@@ -2,34 +2,7 @@
 
 import { useState } from "react";
 import { motion, AnimatePresence } from "framer-motion";
-
-type FAQItem = {
-  question: string;
-  answer: string;
-};
-
-const faqData: FAQItem[] = [
-  {
-    question: "What age groups do you accept?",
-    answer:
-      "We welcome students of all ages, from young children to adults. Our classes are tailored to different skill and age levels.",
-  },
-  {
-    question: "Do I need any experience to join?",
-    answer:
-      "No experience is necessary! We have beginner classes and our instructors guide everyone step by step.",
-  },
-  {
-    question: "Are private classes available?",
-    answer:
-      "Yes, we offer private one-on-one training sessions with our experienced instructors to suit your schedule and learning style.",
-  },
-  {
-    question: "Do you offer trial classes?",
-    answer:
-      "Absolutely! We encourage new students to come try a class to see if our dojo is the right fit for them.",
-  },
-];
+import { faqData } from "../data/faq";
 
 export default function FAQ() {
   const [activeIndex, setActiveIndex] = useState<number | null>(null);
